buildReleaseZip {CommonDataModel}	R Documentation

Create OMOP CDM release zip

Description

First calls buildReleaseZips for given cdmVersions and targetDialects. This writes the ddl sql files to the ddl folder. Then zips all written ddl files into a release zip to given output folder.

Usage

buildReleaseZip(
  cdmVersion,
  targetDialect = listSupportedDialects(),
  outputfolder = file.path(tempdir(), "output")
)

Arguments

cdmVersion
The version of the CDM you are creating, e.g. 5.3, 5.4. Defaults to all supported CDM versions.

targetDialect
The target dialect. Defaults to all supported dialects.

outputfolder
The output folder. Defaults to "output"

Details

If no (or multiple) targetDialect is given, then one zip is written with the files of all supported dialects.

Value

A character string containing the OHDSQL DDL

Examples

## Not run: 
buildReleaseZip(cdmVersion='5.3', targetDialect='sql server', outputfolder='.')

## End(Not run)

[Package CommonDataModel version 0.2.0 Index]