downloadCurrentDdl {CommonDataModel}	R Documentation

Get current DDL sitting on the main branch

Description

Get current DDL sitting on the main branch

Usage

downloadCurrentDdl(
  githubPath = "OHDSI/CommonDataModel",
  pathToCsv = "Sql%20Server/OMOP%20CDM%20sql%20server%20ddl.txt",
  outputFile = paste0("inst/sql/sql_server/OMOP CDM ddl ", Sys.Date(), ".sql")
)

Arguments

githubPath
The path for the GitHub repo containing the package (e.g. 'OHDSI/CommonDataModel').

pathToCsv
The path for the snapshot inside the package.

outputFile
The path where the file should be saved.

Details

This function gets the current ddl on the CDM main branch. It will be taken from the Sql Server folder. The default location is inst/settings/currentOmopDdl.sql.

Value

The current DDL sitting on the main branch of the CommonDataModel repository.

Examples

## Not run: 
downloadCurrentDdl("OHDSI/CommonDataModel",
 pathToCsv="Sql%20Server/OMOP%20CDM%20sql%20server%20ddl.txt")

## End(Not run)

[Package CommonDataModel version 0.2.0 Index]